requirements.txt
aiohttp
anytree
aria2p
asyncio
beautifulsoup4
bencoding
cfscrape
dnspython
feedparser
flask
gdown
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gunicorn
lk21
lxml
pillow
playwright
psutil
pybase64
pyDes
pymediainfo
pymongo
pyrogram
pySmartDL
python-dotenv
python-magic
python-telegram-bot
qbittorrent-api
requests
requests_toolbelt
telegraph
tenacity
tgCrypto
xattr
yt-dlp
